res silenced some more VC2005 compiler warnings and fixed some   compile errors due missing external dependencies.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20012 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/terrain/object/chunklod.cpp
@@ -213,7 +213,7 @@ void csChunkLodTerrainFactory::SetSamplerRegion (
   int resolution = (int)(region.MaxX () - region.MinX ());
   // Make the resolution conform to n^2+1
   resolution = csLog2(resolution);
-  resolution = ((int) pow(2, resolution)) + 1;
+  resolution = (int)((float) pow(2.0f, resolution)) + 1;
 
   hm_x = hm_y = resolution;
   fullsample = terraform->GetSampler (region, resolution);
